+++ NEWS.md
@@ -1,6 +1,7 @@
 Version 0.9.4-dev
 2012-MM-DD
 
+ - Add support for using "mapnik-geometry-type" in filters
  - Enhanced 2.0 to 2.1 style transform:
    - set default marker-placement to line